pyproject.toml
[project]
name = "veriflow"
version = "0.1.0"
description = "Multi-modal content verification framework"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0",
    "tavily-python>=0.5",
    "pymupdf>=1.24",
    "httpx>=0.27",
    "mcp[cli]>=1.0",
    "pydantic>=2.0",
    "pillow>=10.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
